SPlit wirego_plugin code to multiple source files
